Layout cooperation; width, height, size modifiers
The CL updates the layout protocol to permit children measuring
themselves to sizes that do not respect the constraints coming from the
parent. This enables nicer handling for cases when children need to
overflow their parent due to a minimum size they absolutely need in
order to do correct layout and drawing; it also allows more elegant
handling of alignment for children forced by their parents to be larger
than the maximum size they can occupy while doing correct layout and
drawing.
When a child does not respect its constraints, the parent will not find
out about it when doing layout - the size reported by the child will be
automatically coerced inside the incoming constraints before being read
from its corresponding Placeable by the parent. Therefore, the parent
will do layout under the assumption that children will respect their
constraints. Once they assign a position to the child, this will be
automatically offset such that the child will be centered in the space
that the parent assigned to it.
The CL also adds the width, height, size, etc. modifiers, which have
priority over incoming constraints, with conflicts being resolved based
on the protocol explained above. The API shape resembles the already
existing preferredWidth, preferredHeight, etc. modifiers, which will
only affect size when this is possible under the incoming constraints.

Bug: 152608237
Test: ran AndroidLayoutDrawTest, ui-layout tests, new test in LayoutSizeTest
Change-Id: I84939f14525610e9bf5493706895d52a452d425e