Adds defaultFontFamily constructor parameter to Typography
This allows specifying a default family that will be used for provided TextStyles that do not have a family explicitly set.
Additionally this CL adds documentation for all parameters.

Bug: b/148437305
Test: ./gradlew updateApi
Change-Id: I89d07c0730ab21464824754d2f2d17770d6bbdb9
Relnote:"Adds defaultFontFamily constructor parameter to Typography, allowing specifying the default font family that will be used for any provided TextStyles that do not have a family set."