Pump test animation clock before advancing manually
Manually advancing the TestAnimationClock is typically done to
fast-forward an animation, and typically the whole animation. The
problem is that an animation that is started, but hasn't done a frame
yet, doesn't have a start time yet. The start time will be the frame
time of the next frame. If we advance the clock a single time with the
requested milliseconds, the animation will only receive one frame and
take that frame's time as its start time. After advancing, the
animation's play time will thus still be stuck at 0ms.

To solve this, pump the animation clock with 0ms before advancing with
the requested amount. That will set a start time on any animation that
has just been started.

Bug: 152402882
Test: ./gradlew ui:ui-test:cC
Change-Id: I5f3166a460efc621e1e807d59e10819a4d9d4186