Remove StaticDrawer
While there's a material spec of StaticDrawer, it is clear that is actually more for usages showcase than for real app. Real apps are way more complex and usually hybrid, therefore it is better to not confuse people and pollute API surface with something that is not needed, but very easy can be reimplemented (it's just a Box).

RelNote: StaticDrawer was removed. If you need it, use Box with material-spec'ed width instead
Bug: 152764133
Test: StaticDrawer tests removed

Change-Id: I244a7e7878e8cf443c008464229fdeb9a3c24e0f