Deflake ConfigurationActivityTest #2
We are still see occasional flakes specifically on Cuttlefish API 30. This
device seems to be really slow possibly because the branch is still in
development and as a result hasn't been optimised yet. This is a final
attempt to add padding into the test so that we check that the dialog
become visible and stays visible for roughly the expected time before
disappearing. If this isn't successful we should consider the future
of these tests.

Bug: 153057993
Test: ./gradlew wear:wear:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.wear.widget.ConfirmationActivityTest and androidx.wear.widget.ConfirmationActivityTest (from within Android studio against a Cuttlefish AVD created for RVC API 30 [acloudw create --branch git_rvc-release --build-target cf_x86_phone-userdebug], and a Watch round emulator)
Change-Id: I0939833e878811d459abfae2e2732402c3210f08