Isolate expand projection related code
Expand projection makes some assumptions which are bound to go wrong unless
we fully parse the query. On the other hand, optimization part can easily
be achieved by wrapping the query.

This CL tries to limit the access of expand-projection related changes (which
failry complicated the parser, necessarily).

For the isolation, this CL reverts the SQLParser and introduces
ExpandableSQLParser for backwards compatibility.
The Query rewriting logic now goes through a QueryRewriter API that we
can eventually swap with the new rewriter.

Even though this CL duplicates code, I thought it is better so that we
don't carry over extra information for ParsedQueries. The downside is
though, if expand projection is on, we'll be parsing the queries twice.

Bug: 153387066
Test: existing tests are passing.
Test: kotlin and java tests now run w/ and w/o expand projection

Change-Id: I28c5370d0eb97b80e0498b7c4783032038e892f1