Fix the DND crash when CameraX initialization and open camera.
The crash happens if the following conditions are met.
* An early Android P(28) release
* Camera hardware level is INFO_SUPPORTED_HARDWARE_LEVEL_LEGACY
* Device Do Not Disturb (DND) mode is enabled

It will crash by 2 scenarios,
(1) Enable DND mode and trigger CameraX initialization, it will crash during initialization.
(2) After CameraX initialized, enable DND mode and bind use cases, it will crash when try to open camera.

The CL tries to not crash for both scenarios.
For scenario 1, it set a specific exception CameraUnavailableException to the initialization ListenableFuture to notify the user that it is caused by enabling DND mode.

The CL contains
(1) Create checked exception CameraUnavailableException with error code and InitializationException.
(2) Wrap the DND fail exception (CameraAccessExceptionCompat) by the CameraUnavailableException. Set the exception to InitializationException and initialization ListenableFuture instead of crashing the application.
(3) Reset the Camera2CameraImpl internal state when encountering DND fail instead of crashing.

Relnote: "Fix the crash on app startup when initialising CameraX while phone is in Do Not Disturb mode. An InitializationException contains a CameraUnavailableException will be set to the ListenableFuture of the intialization result instead of crashing the application."

Bug: 149413835
Test: ./gradlew camera:camera-camera2:connectedAndroidTest; Simulate the DND fail case by modifying CameraManagerCompat and manually test
Change-Id: I9909af08a8305fcf8e603d709d98b4488f1af69f