Update SelectionContainer to calculate the handle positions in onPositioned callback
Previously we were calculating the coordinates during the composition which is not always correct as it is possible that during this composition we already updated some layouts or their modifiers which dirties the measured sizes/positions. Instead we should react on the next onPositioned callback, recalculate the updated handles positions and then repositioning will happen automatically as we use State. I also added remember for the modifier being applied to SelectionContainer to optimize it a bit as every time we create a new modifier and apply it it will require some unnecessary remeasure work as we know this modifiers are not changing in fact.

Bug: 153568460
Bug: 147762988
Bug: 152594949
Test: manually, global coordinates are calculated correctly now if we resolve them inside SelectionManager.updateHandleOffsets()
Change-Id: Ib33cef8d6bf267b6f01b47be910d95fc7388ff54