Add framework for sending accessibility events
We save the old semantics tree and compare it with the new semantics
tree. For property changes, we send out correspsonding events. For
structural changes, we send out the sub tree change events. This CL adds
the logic in dispatchDraw. Will add semantics change notification
function later and move the logic here into it in a separate CL.

Test: tested with the demo app. The check/unchecked, selected/unselected annonuces upon click.
Also tested with progress indicator demo.

Change-Id: I1a5266a205505d1b673520511ed14025a1b188d2