Add ViewTreeSavedStateRegistryOwner
Add utilities for setting and querying the SavedStateRegistryOwner for a view subtree.
This is the SavedStateRegistryOwner that is responsible for managing the saved state
for the views within the hierarchy. The ViewTreeSavedStateRegistryOwner for a view
that does not have a SavedStateRegistryOwner set explicitly is the SavedStateRegistryOwner
of its parent.

This mirrors the API in ViewTreeLifecycleOwner.

Test: new tests
Change-Id: I1bfffdf7fb06ec258c5329107c901db3dd5a1fac