Support alignment lines reads from LayoutModifier
The CL adds support for reading alignment lines in the measure and
positioning blocks of LayoutModifier. Although the API of LayoutModifier
was already supporting alignment lines reads, trying to do so was
resulting in a crash or buggy values due to missing wire up. Note that
providing an alignment line in LayoutModifier is still unsupported and
will be fixed in a further CL.

Relnote: Fixed alignment lines reads from LayoutModifier.
Bug: 155736449
Test: AndroidLayoutDrawTest

Change-Id: I9dd1dbeded5f7d6e3d0b7f50a09fb1b0618ca769