Ensure we disptachOnDestinationChanged when graph restored
When the nav graph is restored, since we already have the proper
backstack, we never make a call to navigate(). That means that we do not
disptach to any OnDestinationChanged listeners that may have been
registered on the restored nav graph.

If a graph is created and the backstack is already rebuilt, lets
dispatch to any OnDestinationChangedListeners that are added.

Test: NavControllerTest
Bug: 155218371
Change-Id: Ie50e57cc5fbacbde17af2e5eaca7cc1fdeecacba