MediaRouter: Create MediaRoute2Provider
As a first step for using framework MediaRouter2 more actively,
this CL introduces a class MediaRoute2Provider which provides
routes and RouteControllers by using framework MediaRouter2.

The new logic will run only when sGlobal.mUseMr2FwkActively is true,
which should be manually done for testing until the implementation
of the new logic is complete. (Also enableTransfer() is needed.)

Bug: 155378155
Test: Manually checked the logs of the DiscoveryRequest inside of
      MediaRoute2Provider#updateMr2FwkCallbacks().
Change-Id: Ia31a08ecd15d9500a97237c0ed03e8a366f00f0f