Cache SemanticsNode and InputDispatcher in doGesture()
The cached value lives for the duration of the do(Partial)Gesture method
in Actions.kt. When the scope is created, the SemanticsNode is fetched
immediately, requiring a single synchronization step, and the
InputDispatcher is created from it. Both can now be used without any
further synchronization steps.

To prevent leaking the node or the input dispatcher, they are discarded
after the gesture lambda that was passed to doGesture() or
doPartialGesture() has executed.

Bug: 156591919
Test: ./gradlew ui:ui-test:cC
Change-Id: If79e176960b6da68abb5021bd2c60e2edbed0649