API Change: make UseCaseGroup public and experimental
Background:

UseCaseGroup is a grouping mechanism for UseCases so they will affect each other in terms of crop rect and other. It also gives the group member share properties like scale type and aspect ratio. The feature is a popular request from users. It's also used by higher level API like the view artifcat.

This CL:

The feature is currently an internal feature. This CL makes it public and marks it as an experimental API.

Bug: 153096869
Test: manual test and ./gradlew bOS
Change-Id: Ia5fd9eb47707b3c61b367b7adab7f4eb93005d31