Change AtomicFile to use rename-into-place.
The previous implementation of backing up beforehand doesn't handle
the case where the file is created for the first time, and might leave
a corrupted file in case of failure.

This new implementation creates a new file for writing data into, and
renames it into the place of the original file after writing
finished.

Bug: 151959443
Test: ./gradlew core:core:connectedCheck
      -Pandroid.testInstrumentationRunnerArguments.class=androidx.core.util.AtomicFileTest
Change-Id: If39befe800727078252be04beba5cfe90f3dc205