Move AndroidOwner interface into a separate file (Step 1/2)
Lets extract AndroidOwner into its own file to keep AndroidComposeView small and simple.
To not lose the git history I do it in two cls, first rename the AndroidOwner file to AndroidComposeView and then create AndroidOwner file again and more this interface there. If we do it in one step huge AndroidComposeView will be displayed as a new file.

Bug: 156889291
Test: manually, just rename
Change-Id: I48bd2af2731588bbe7a0086e96149829dc451c93