Support RTL LayoutDirection in ViewPort
The definition of "start" and "end" in scaleType depends on the LayoutDirection of the Preivew view. For {@link LayoutDirection#LTR}, the start is (0,0) and the end is (width, height); for {@link LayoutDirection#RTL}, the start is (width, 0) and the end is (0, height). The output rect needs to be transformed to match View properties.

- Update ViewPort to enable the LayoutDirection property
- Update ViewPorts to include LayoutDirection in calculation. If it's RTL, flip the output rect vertically.
- Move ViewPort related tests to its own class and added new test for LayoutOrientation.

Also removed the unused CameraX#getRotatedUseCaseSizes and its tests.

Bug: 153096869
Test: manual test and ./gradlew bOS
Change-Id: Ibd127a3d9a6ab98e792b8bb993606b3f94c9b2a9