Attempt to reenable test that was failing on cuttlefish.
The test was causing process crashes on CuttleFish API 29
devices that were assumed to not be an issue with RecyclerView
or this test.
It was temporarily suppressed all sdk versions 29 and above until
the issues were resolved so that RecyclerView CLs were not blocked.

Attempt to reenable in hopes that they are ok now.

Fixes: 142682722
Test: Tests not run, want to see that they pass in our testing infrastructure.

Change-Id: Iaf97c4293333bfcba622b082773d3bec891fc38e