Reset references in solver State on #reset()
In ConstraintLayout we are reusing the Measurer and
ConstraintWidgetContainer root instance across recompositions. However,
before this CL we were not clearing properly the children references
between measurements, which was causing a crash because of trying to
measure obsolete detached LayoutNodes.

Relnote: Fixed a ConstraintLayout bug causing a crash on recompositions.
Bug: 158164341
Bug: 157638113
Test: ConstraintLayoutTest
Change-Id: Ibee5afd640816add9bd0c1545eee2ef747e9c2d2