Generate and use a mapping of shared element names
The whole point of shared element transitions is to
match up views from the firstOut Fragment with the
lastIn Fragment. In order to support missing Views and
a 1:1 matching of Views, we need to build a mapping
from the firstOut Views to the lastIn Views.

This mapping allows us to exclude Views that don't
exist in both Fragments as well as exclude shared
element transitions entirely that have no Views that
are shared (i.e., an empty mapping).

By applying this mapping, we can correctly swap
from the set of outgoing firstOut shared element Views
to the incoming lastIn shared element Views at the
appropriate times relative to beginDelayedTransition().

Test: considerably more shared element tests pass
BUG: 147749580
Change-Id: Id9159c510fd9af51954abc6d5564f9729089a736