Try to de-flake FlowAsLiveDataIntegrationTest
This test sometimes times out in CI (30%) even though isolated Forrest
runs do not flake (20 runs).

It could possibly flake if another test is leaving arch executors in
a bad state. This CL adds a before action to clear any set delegate.

I've also slightly modified the test to finish the custom flow without
waiting for the timeout so that this test itself also do not leak a
coroutine.

Bug: 159212029
Test: FlowAsLiveDataIntegrationTest
Change-Id: I8ec2c08136ba954baa0c7147b1282cfb32f9fde2