Start ActivityResultRegistry request codes at 0x10000
Rather than start the request codes at the exact upper
bound of the lower bits (0xFFFF), update the starting
request code to always be within the upper bit range
(i.e., starting exactly one request code higher than
before).

This ensures that there is no overlap, not even in a single
bit, with the request codes previously valid for use with
startActivityForResult in an Activity.

Test: Ran integration testapp
Change-Id: I7b668edde5f95e262f010cab21106ffa095a2632