Fix crash in nested interop
Before this CL, having Compose inside Views inside Compose would lead to
a crash during the outer composition disposal. This was happening due to
the inner AndroidComposeView being onDetachedFromWindow() twice, which
was in turns happening because of the parent AndroidViewHolder removing
it during its own onDetachedFromWindow(): so the AndroidComposeView was
onDetachedFromWindow once during its removal from the parent and once
because the detaching dispatching was continuing from the parent. This
CL fixes this behavior by only removing it after the AndroidViewHolder
is detached, such that, when removed, it will not be detached again due
to it being marked as already detached.

Test: ComposedViewTest
Bug: 159106722
Change-Id: I0410e073064a84165b7b3663e14cb5e0f9ccf81c