Add asMutableList() to MutableVector
Relnote: "Added asMutableList() to MutableVector to allow it
to be passed to public API without having to copy the entire
list."

I did tests to see if MutableVector could implement
MutableList, and it turns out that even though "forEach(),"
for example, is an instance method, the List extension
function overrides the instance version. This makes it
always fail to use the high-performance version of
common methods and the performance (and allocations)
suffer.

Test: new tests
Change-Id: I298df3722ef0fa4bcce7cc0398b931ee934bb233