Fix compose testing not timing out properly.
ComposeIdlingResource was quite aggresively subscribing into Handler
queue which was blocking any other events to go though (e.g. we would
subscribe 200k times in few seconds). I added a delay to post every 20
millis which unblocked the error pipeline.

Also the error reporting from Espresso was not mentioning what exactly
is the issue when we time out. So I have added our infra that catches
the Expcetion from Espresso and provides more info if we have it.

Also added quite good test coverage for this case. Note that the
original Handler queue flooding is not somthing that even Junit timeouts
would be able to tackle so if we regress it will have to be some global
runner deadline that kills the tests.

Bug: b/159105007
Test: Added
Relnote: N/A
Change-Id: Ie4acb853c806ccf81d1ad6f764f8fe8d0aa820a0