Adds script to aid androidx.ui -> androidx.compose migration
This temporary script can be removed after the migration is complete.

This script assumes that the current CL migrates a library (or multiple), with the previous Cl containing the library in its pre-migrated state.
The script then runs createArchive on both CLs, and diffs the output, as well as providing a japi-compliance-checker command to diff the actual jar inside each aar.

Bug: b/160233169
Test: ../development/validateComposeModuleMigration.sh androidx.ui:ui-material androidx.compose.material:material --build
Change-Id: Ib85bc274729b4aa41cd1e3db21ad0c99ab93d220