Fix incorrect buffer copy in HdrImageCaptureExtenderImpl example
The sample buffer was simply copying the input YUV buffers into the
output YUV buffers making the assumptions that the input and output
planes have matching strides. Instead this has been changed to use the
strides for the planes.

This does not try to optimize by doing bulk copy if the strides are
smaller.

Bug: b/159210415
Test: androidx.camera.extension.ExtensionTest
Change-Id: I61b1dce1aba4e4ea46e8a1487f1f2a791c55f08e