Use unclipped bounds when asserting node bounds
When asserting bounds and positions of nodes, unclipped bounds are
preferred over clipped bounds. Otherwise you won't be able to
distinguish between a correctly positioned node that is clipped to
certain bounds, and an incorrectly positioned node that happens to be
clipped to the same bounds.

Add positionInRoot to the SemanticsNode API and use the unclipped bounds
in BoundsAssertions.

Fixes: 161336532
Test: ./gradlew ui:ui-test:cC
Relnote: "Added `SemanticsNode.positionInRoot` to get the position of a
SemanticsNode relative to the root of the Compose hierarchy"

Change-Id: Icdf262558341d6441ae91dfb8807d19f88f5b3fb