Add paging-rxjava3 artifact
This adds an rxjava3 extensions artifact for paging v3. It's ported from the rx2 artifact, but with all the deprecated APIs from that artifact removed.

I had to update the coroutines version to 1.3.7 (technically 1.3.6 was when it was introduced, but with a serious issue affecting android lint that is fixed in 1.3.7). I don't know how to get those prebuilts updates included in my CL though, not sure if it's necessary. Let me know.

RelNote: "Add RxJava3 compatibility wrappers for PagingSource and
RemoteMediator"
Fixes: 161480176
Test: ./gradlew paging:paging-rxjava3:test

Change-Id: I49ef38fde9b84f92dd272ed0b1cd1719fbbf1761