Ensure main thread in LifecycleRegistry methods.
RelNote: "LifecycleRegistry now verifies that its methods are called
on main thread. It was always a requirement for lifecycles of
activities, fragments etc. An addition of observers from non-main threads
resulted in hard to catch crashes in runtime. For LifecycleRegistry
objects that owned by your own components, you can explicitly opt out
from checks by using LifecycleRegistry.createUnsafe(...), but then you
have to ensure that a proper synchronization is in place when
this LifecycleRegistry is accessed from different threads"

bug: 137392809
Test: LifecycleRegistryTest.java
Change-Id: Ie72809b34d4e66c20b9694548359a1209d438f3d