[Text Field] Calculate TextDirections After Layout Again.
Before this CL, the TextDirection of the Selection Handles are calculated after the Selection is updated but before Laying out the text again. This causes crash when pasting longer text to the end of the text, because the selection end is updated, but it's out of range of the current text layout.
The correct place to calculate the TextDirection is just before using it, aka, drawing the handles.

Bug: 139311556
Test: Manually tested.
Test: ./gradlew :ui:ui-text:testDebugUnitTest
Change-Id: Iec791650aa66e654547f53bc8bb261d07f049d1b