Add parameters to Layout Inspector API.
Support an initial limited set of parameter values to be included
with the Layout Inspector API. Parameters with values that are not
currently supported are currently not included.

Also add a generated id to all nodes that are not render nodes.
This allows the agent to identify parameters to a node such that
the parameters can be sent in a separate message.

Relnote: N/A
Bug: 158130075
Test: Unit test added

Change-Id: Ic0dd15b80a445078e00f0b1003a086863c9dd93c