Add an ExperimentalFocus annotation
The Focus API is experimental and is likely to change in the future. Add an experimental annotation that can be added to all focus related API. Users will have to @OptIn to use focus. We will remove the annotation when the focus API is stable.

Bug: 160820961
Test: This annotation is not yet used, so no applicable tests.
Relnote: Added ExperimentalFocus which is an @OptIn annotation for Focus API.
Change-Id: I1544bc717d24434ccc2295244ddb6bf7b2dde3c8