Cancel async PageSwiperManual after tests
If we don't, it will continue injecting MotionEvents to the
Instrumentation instance when the test is already finished. Those
MotionEvents will now be delivered to whatever application comes in the
foreground after the test (e.g. the launcher), which leads to a
SecurityException. When that happens, the whole test suite can be taken
down in certain circumstances.

Fixed by separating the instantiation and execution of the
ManualSwipeInjector into separate statements, and adding a cancel
method to both ManualSwipeInjector and PageSwiperManual that can be
called from another thread.

Bug: 160590776
Test: ./gradlew viewpager2:viewpager2:cC \
        -Pandroid.testInstrumentationRunnerArguments.class=\
        androidx.viewpager2.widget.FakeDragTest
Change-Id: I8dabea83b14828a51a69f42a8e618bb738bf7018