Make LinearLayoutCompat.LayoutParams extend from the platform's LinearLayout.LayoutParams
The current one extends directly from ViewGroup.MarginLayoutParams,
however it's unable to override a crucial method that is private to
platform and that causes views created programatically to be missing
the weight attribute. By extending directly from
LinearLayout.LayourParams, we get that class' overridden
implementation of the method
(protected void encodeProperties(@NonNull ViewHierarchyEncoder encoder))
which should be enough to solve the issue.

Bug: 140349060
Relnote: N/A
Test: none

Change-Id: I8774f6dca5f7176307d571ea267e3d62c402092b