Revert "Revert "Ensure main thread in LifecycleRegistry methods.""
This reverts commit f20bbf070d328ea5d988c9edc4b8a12dda17364e.

RelNote: "LifecycleRegistry now verifies that its methods are called
on main thread. It was always a requirement for lifecycles of
activities, fragments etc. An addition of observers from non-main threads
resulted in hard to catch crashes in runtime. For LifecycleRegistry
objects that owned by your own components, you can explicitly opt out
from checks by using LifecycleRegistry.createUnsafe(...), but then you
have to ensure that a proper synchronization is in place when
this LifecycleRegistry is accessed from different threads"

bug: 137392809
Test: LifecycleRegistryTest + existent tests pass
Change-Id: I287945e56b9404efa761a538f9d4f5a5ddd9ba98