Improve ./gradlew studio for playground
Previously, properties files were not set properly for playground tasks
as they were not accounting for the right path.

I've also changed playground's settings file to validate the JVM version
before continuing. If invoked with the wrong JVM (.e.g. 8), it will
fail with cryptic errors so we are better off detecting it and failing
gracefully. Now we tell developer what to do if JDK is wrong.

Bug: n/a
Test: local
Change-Id: Ifeb24635b246b9b876d6194f32205d958dd441c7