A customized TabLayout to be used as top navigation bar in leanback
Using viewpager instead of viewpager2 as there are some with focus flow
when DPAD_DOWN is pressed while focus is on tablayout. The focus does
not flow from the viewpager2/underlying recyclerview to the contents of
viewpager2.

Test: Unit test and added sample demo

Relnote: "Adding a customized TabLayout to be used as top navigation bar
in leanback"

Change-Id: I1e3048d300106ad0b0f9ef55f14e7070793a579a